refactor: split runner r2 cache module
