+++ crates/runner/src/r2_cache/mod.rs
@@ -1,0 +1,206 @@
+//! R2 cache for `runner build` template artifacts.
+//!
+//! The current shared cache stores reusable template objects at
+//! `runner-templates/{template_hash}.tar.zst` in the existing
+//! `R2_USER_STORAGES_BUCKET_NAME` bucket. Rootfs images are customized locally
+//! because they contain guest binaries and host-local CA material.
+//! Snapshot files are always created locally because they contain host-specific
+//! state (page cache, kernel metadata).
+//!
+//! ## Lifecycle
+//!
+//! 1. `runner build` computes a `template_hash` for the shared R2 object and
+//!    a separate `rootfs_hash` for local images.
+//! 2. `--warm-rootfs-cache` ensures the template R2 object exists. Existing
+//!    objects are checked with HEAD only; normal builds still validate the
+//!    archive before using it.
+//! 3. Normal builds materialize the template object into a per-attempt local
+//!    file (or build/upload it on miss), move the verified template into
+//!    `rootfs.ext4.staging`, customize the staging image locally, verify it,
+//!    and then atomically commit the rootfs.
+//!
+//! Atomicity guarantees:
+//! - Multipart upload is atomic from consumer POV (object only appears after
+//!   `CompleteMultipartUpload`); abandoned segments are auto-cleaned by R2's
+//!   default 7-day lifecycle.
+//! - Template download unpacks into a sibling staging directory and only renames
+//!   `template.ext4` into the caller's destination after the archive is fully
+//!   decoded and validated.
+//!
+//! Configuration semantics: `from_env` returns `Ok(None)` only when **all four**
+//! `R2_*` env vars are unset or empty (dev/test path). Setting 1-3 of 4 is a
+//! fatal `PartialConfig` error — almost certainly a typo'd secret rotation, and
+//! silently disabling cache fleet-wide is worse than failing the deploy.
+//!
+//! Streaming: upload avoids temp files by using a `tokio::io::duplex` pipe to
+//! couple the sync tar+zstd producer (on a blocking thread) to the async
+//! multipart consumer. Download streams the S3 body through `SyncIoBridge` into
+//! a sibling staging directory, then renames the extracted `template.ext4` to
+//! the caller's destination. Callers that coordinate shared output paths should
+//! pass an attempt-scoped destination and perform their own final publish step.
+//! Memory peak per upload ≈ `(2 + CONCURRENCY + 1) × PART_SIZE` — duplex buffer,
+//! in-flight upload chunks, and the part being read — bounded regardless of
+//! image size. Currently ~112 MiB with `PART_SIZE` = 16 MiB and `CONCURRENCY` = 4.
+//!
+//! Image size limit: `PART_SIZE * 10000 ≈ 160 GiB` (S3 multipart hard limit).
+//! Current images are well under 30 GiB; revisit if `PART_SIZE` decreases.
+//!
+//! ## R2-side cleanup
+//!
+//! Completed objects are **never deleted on upload**. Each template cache version
+//! bump or template build script change produces a new template hash and
+//! orphans the previous object.
+//!
+//! Cleanup happens via `gc_older_than`, called from `runner gc` (which the
+//! deploy playbook runs after every release). Default TTL is 7 days. Each
+//! host runs the same scan independently — `DeleteObjects` is idempotent for
+//! already-absent keys, so concurrent fleet execution is safe and costs
+//! ~1 LIST + 1 batched DELETE per host per gc cycle.
+//!
+//! R2's default 7-day lifecycle rule only cleans abandoned multipart
+//! segments, **not** completed objects — which is why we need our own scan.
+//!
+//! **Clock skew caveat**: `gc_older_than` uses local `SystemTime::now()` to
+//! compute the cutoff. If the host clock drifts ahead of R2 server time by
+//! more than the TTL, GC over-deletes (worst case: wipes everything older
+//! than `now_local - keep_days`, even objects that were just uploaded by
+//! peers with correct clocks). Mitigation: keep NTP healthy. A clock behind
+//! R2 is the safe direction (under-deletes, no data loss).
+//!
+//! ## Cancellation safety
+//!
+//! All operations in this module are safe to cancel (drop the future) at any
+//! await point — no permanent state is left in an inconsistent way:
+//!
+//! - **Local staging directory**: a hard-killed template download may leave a
+//!   `*.download.tmp/` directory beside the destination. The next download
+//!   removes it as the first action, so the leak is bounded to one stale dir
+//!   per destination and self-heals on next attempt.
+//! - **R2 multipart upload session**: once `create_multipart_upload` returns,
+//!   an owned guard aborts the upload on normal errors and schedules a
+//!   best-effort abort if the upload future is cancelled before disarm.
+//! - **`spawn_blocking` pack / unpack tasks**: tokio cannot cancel
+//!   blocking tasks. After parent cancellation, the producer/consumer
+//!   thread runs until it hits BrokenPipe or natural EOF — wasted CPU for
+//!   a few seconds, no resource leak.
+//!
+//! ## Corrupt-object eviction
+//!
+//! A structurally-valid archive whose extracted content lacks template.ext4
+//! (e.g. uploaded by an old/buggy producer, or attacker-controlled IAM
+//! key writing a bogus tar to a predicted hash key) would otherwise
+//! dead-lock the fleet's cache for that hash: every host downloads → unpacks
+//! → finds no template → rebuilds locally → dedup-skips upload because the bad
+//! object already exists.
+//!
+//! `cmd::build::run_build` defends by passing `force = true` to upload
+//! whenever template download classifies an object as invalid. That bypasses the
+//! dedup check and atomically overwrites the bad object via multipart complete.
+//!
+//! ## Tar entry security
+//!
+//! The `tar` crate (0.4) has two relevant behaviors when consuming an
+//! attacker-influenced archive:
+//!
+//! 1. **Path traversal (`..` components) is silently dropped**. Verified by
+//!    `unpack_rejects_path_traversal`. The malicious entry is skipped; the
+//!    staging dir ends up missing template.ext4; the template download helper
+//!    rejects that as an invalid object and the caller rebuilds locally. Safe.
+//!
+//! 2. **Symlink and hardlink entries are rejected**. `unpack_from_reader`
+//!    iterates entries and rejects any whose type is not `Regular`,
+//!    `Continuous`, or `GNUSparse` — symlinks, hardlinks, character/block
+//!    devices, FIFOs, and extended-header pseudo-entries all cause an
+//!    immediate error, preventing an attacker with R2 write access from
+//!    crafting a tar where expected filenames are symlinks to host paths.
+//!    (`GNUSparse` is retained for forward compatibility with any future
+//!    sparse file in the archive; template.ext4 itself is packed as a
+//!    regular file.)
+//!
+//! **Maintenance note**: `try_download_template_file_by_key` verifies that the
+//! archive contains a regular `template.ext4` file and classifies a missing or
+//! non-file member as an invalid object. If you add a new required member to the
+//! template R2 archive, extend that validation accordingly — otherwise an
+//! attacker-controlled tar that omits the new file would go undetected.
+
+use aws_sdk_s3::error::SdkError;
+
+mod archive;
+mod config;
+mod download;
+mod gc;
+mod keys;
+mod multipart;
+mod upload;
+
+#[cfg(test)]
+mod tests;
+
+#[derive(Debug, thiserror::Error)]
+pub enum R2Error {
+    #[error("R2 partially configured ({}/4 set), missing: {}", present.len(), missing.join(", "))]
+    PartialConfig {
+        present: Vec<String>,
+        missing: Vec<String>,
+    },
+    #[error("s3: {0}")]
+    S3(String),
+    #[error("io: {0}")]
+    Io(#[from] std::io::Error),
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum R2DownloadError {
+    #[error("request failed: {0}")]
+    Request(#[source] R2Error),
+    #[error("invalid cache object: {0}")]
+    InvalidObject(#[source] R2Error),
+    #[error("local filesystem failed: {0}")]
+    Local(#[source] R2Error),
+}
+
+impl R2DownloadError {
+    pub fn is_invalid_object(&self) -> bool {
+        matches!(self, Self::InvalidObject(_))
+    }
+}
+
+impl<E, R> From<SdkError<E, R>> for R2Error
+where
+    E: std::fmt::Debug,
+    R: std::fmt::Debug,
+{
+    fn from(e: SdkError<E, R>) -> Self {
+        Self::S3(format!("{e:?}"))
+    }
+}
+
+/// Cache handle. Cheap to clone (the underlying SDK client is `Arc`-internal).
+#[derive(Clone)]
+pub struct R2ImageCache {
+    client: aws_sdk_s3::Client,
+    bucket: String,
+}
+
+impl std::fmt::Debug for R2ImageCache {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("R2ImageCache")
+            .field("bucket", &self.bucket)
+            .finish_non_exhaustive()
+    }
+}
+
+pub(super) fn io_other<E: std::fmt::Display>(e: E) -> std::io::Error {
+    std::io::Error::other(e.to_string())
+}
+
+#[cfg(test)]
+impl R2ImageCache {
+    /// Test-only constructor. Lets unit tests inject a mock `aws_sdk_s3::Client`
+    /// (built via `aws_smithy_mocks::mock_client!`) without going through
+    /// `from_env`, which reads process env vars. Production code MUST construct
+    /// via `from_env`.
+    pub(crate) fn with_client(client: aws_sdk_s3::Client, bucket: String) -> Self {
+        Self { client, bucket }
+    }
+}
